requirements.txt
django == 3.1.1
djangorestframework
django-cors-headers
asgiref==3.2.10
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
Pillow==5.0.0